WebSockets (#5)
This PR adds websockets support to the project, dividing the code into a client (ui) / server (simulator) architecture.

It also adds a cli interface for driving both the server and the ui.

A lot of cruft from the earlier phases of this project was also deleted.

Docs have been updated.